+++ src/trac_ik/trac_ik_kinematics_plugin/CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog for package trac_ik_kinematics_plugin
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.1 (2024-04-12)
+------------------
+
+2.0.0 (2024-03-29)
+------------------
+* hide parameters header file, as it is implemetation
+* remove boost from trac_ik_lib and trac_ik_kinematics_plugin
+* Set default solve type to Distance
+* Contributors: Ana C. Huaman Quispe, Kenji Brameld
+
 1.6.6 (2021-05-05)
 ------------------
 * removed bad depends